Have done a lot over the past few days. I have gotten a new endpoint setup to test getting data from the database rather than getting the 1 JSON blob to use on the frontend. It has worked. All data now gets updated at the end of the graph to use in this new model. Worked on the chat feature getting chats and context back and forth. Tested via command line, no frontend implementation yet. New databases for model portfolio stats. New extraction agent to parse uploaded portfolio files instead of having to enter positions every single time. Overall all, coming along nicely

+++ backend/app/models/position.py
@@ -1,6 +1,6 @@
 from pydantic import BaseModel
 class Position(BaseModel):
     ticker: str
-    shares: int
+    shares: float
     costBasis: float
     currentBasis: float
